+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.26] - 2026-01-10
+
+### Fixed
+
+- Fixed issue with missing refreshes
+
+### Added
+
+- Added Target lines, and Additional lines, to settings
+
 ## [0.5.25] - 2026-01-09
 
 ### Added
@@ -185,6 +195,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release. This document will be updated for subsequent releases.
 
+[0.5.26]: https://github.com/batrachianai/toad/compare/v0.5.25...v0.5.26
 [0.5.24]: https://github.com/batrachianai/toad/compare/v0.5.23...v0.5.24
 [0.5.23]: https://github.com/batrachianai/toad/compare/v0.5.22...v0.5.23
 [0.5.22]: https://github.com/batrachianai/toad/compare/v0.5.21...v0.5.22